Update Swagger definitions, deps, and library components
