+++ .eslintrc
@@ -42,6 +42,7 @@
     "import/extensions": 0,
     "import/prefer-default-export": 0,
     "@typescript-eslint/camelcase": 0,
-    "@typescript-eslint/no-explicit-any": 0
+    "@typescript-eslint/no-explicit-any": 0,
+    "no-restricted-exports": 0
   }
 }
