+++ README.md
@@ -1,18 +1,18 @@
 <p align="center">
-  <img height="60" src="https://user-images.githubusercontent.com/7189823/133838642-9a05e1ec-9a79-46ae-b22e-a8b931caf233.png" alt="Shortcut logo">
+  <img height="60" src="https://github.com/useshortcut/shortcut-client-js/assets/739719/b1a2efa0-6de1-4bc4-8525-563c53a9b2ba" alt="Shortcut" />
 </p>
 
 <h3 align="center">
-  @useshortcut/client
+  @shortcut/client
 </h3>
 
 <p align="center">
-  A library for interacting with the Shortcut (formerly Clubhouse) REST API
+  A library for interacting with the Shortcut REST API.
 </p>
 
 <p align="center">
-  <a href="https://www.npmjs.org/package/@useshortcut/client">
-    <img src="https://badge.fury.io/js/@useshortcut%2Fclient.svg" alt="Current npm package version." />
+  <a href="https://www.npmjs.org/package/@shortcut/client">
+    <img src="https://badge.fury.io/js/@shortcut%2Fclient.svg" alt="Current npm package version." />
   </a>
   <a href="https://circleci.com/gh/useshortcut/shortcut-client-js">
     <img src="https://circleci.com/gh/useshortcut/shortcut-client-js.svg?style=shield" alt="Current CircleCI build status." />
@@ -21,7 +21,7 @@
     <img src="https://img.shields.io/badge/PRs-welcome-brightgreen.svg" alt="PRs welcome!" />
   </a>
   <a href="https://twitter.com/intent/follow?screen_name=useshortcut">
-    <img src="https://img.shields.io/twitter/follow/useshortcut.svg?label=Follow%20@useshortcut" alt="Follow @useshortcut" />
+    <img src="https://img.shields.io/twitter/follow/useshortcut.svg?label=Follow%20@shortcut" alt="Follow @shortcut" />
   </a>
 </p>
 
@@ -31,17 +31,19 @@
 
 ### Installation
 
-You can install `@useshortcut/client` with [NPM](https://www.npmjs.com/) or [Yarn](https://yarnpkg.com).
+You can install `@shortcut/client` with [NPM](https://www.npmjs.com/) or [Yarn](https://yarnpkg.com).
 
 ```shell
-npm install --save @useshortcut/client
+npm install --save @shortcut/client
 ```
 
 or
 
 ```shell
-yarn add @useshortcut/client
+yarn add @shortcut/client
 ```
+
+> Our legacy organization name is `@useshortcut`. While this will still contain copies of all future released package versions, we **strongly** recommended switching to our primary `@shortcut` organization namespace.
 
 ### How to Get an API Token
 
@@ -62,8 +64,8 @@ Requests made with a missing or invalid token will get a `401 Unauthorized` resp
 To see all available exports, take a look at the [API documentation](https://useshortcut.github.io/shortcut-client-js/) or check out the `.d.ts` files in this repository.
 
 ```javascript
-import { ShortcutClient } from '@useshortcut/client';
-// const { ShortcutClient } = require('@useshortcut/client');
+import { ShortcutClient } from '@shortcut/client';
+// const { ShortcutClient } = require('@shortcut/client');
 
 const shortcut = new ShortcutClient('YOUR_API_TOKEN'); // See https://github.com/useshortcut/shortcut-client-js#how-to-get-an-api-token
 
@@ -74,7 +76,7 @@ shortcut.listProjects().then((response) => console.log(response?.data));
 
 ## Play with It
 
-You can play with it in your web browser with this live playground: 
+You can play with it in your web browser with this live playground:
 
 - [CodeSandbox](https://codesandbox.io/s/useshortcut-client-playground-48kq1)
 
@@ -82,4 +84,4 @@ You can play with it in your web browser with this live playground:
 
 [Documentation for this client](https://useshortcut.github.io/shortcut-client-js/).
 
-[Documentation for the REST API](https://shortcut.com/api/rest).
+[Documentation for the REST API](https://developer.shortcut.com/api/rest/v3).
