+++ docker-compose.yml
@@ -1,0 +1,11 @@
+version: '3.8'
+
+services:
+  shortcut-client-js:
+    container_name: shortcut-client-js
+    image: node:20-slim
+    stdin_open: true
+    tty: true
+    volumes:
+      - .:/shortcut-client-js
+    working_dir: /shortcut-client-js
