+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@useshortcut/client",
-  "version": "1.0.0-beta.1",
-  "description": "A Promise based library to the Shortcut (formerly Clubhouse) REST API",
+  "name": "@shortcut/client",
+  "version": "1.1.0",
+  "description": "A Promise based library to the Shortcut REST API",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
   "files": [
@@ -18,43 +18,46 @@
   },
   "license": "MIT",
   "scripts": {
-    "build": "yarn build:client && tsc",
-    "build:client": "swagger-typescript-api -p ./schema/shortcut.swagger.json -o ./src/generated --clean-output --axios --modular --templates ./templates && yarn add-typedoc-comments",
+    "build": "yarn build:client && yarn build:add-typedoc-comments && tsc",
+    "build:client": "swagger-typescript-api -p ./schema/shortcut.swagger.json -o ./src/generated --clean-output --axios --modular",
+    "build:add-typedoc-comments": "npx jscodeshift -t scripts/add-typedoc-comments.ts --extensions=ts --parser=ts src/generated/**",
     "build:docs": "typedoc --entryPointStrategy expand ./src --exclude 'src/__tests__/**'",
-    "sync:schema": "wget https://shortcut.com/api/rest/v3/shortcut.swagger.json --destination ./schema/shortcut.swagger.json && yarn validate:schema",
-    "validate:schema": "swagger-cli validate ./schema/shortcut.swagger.json",
     "validate:examples": "npx tsc examples/*.ts --noEmit",
-    "add-typedoc-comments": "node ./node_modules/.bin/jscodeshift -t scripts/add-typedoc-comments-transform.js src/generated/**",
+    "prepublishOnly": "yarn build",
+    "sync:schema": "wget https://developer.shortcut.com/api/rest/v3/shortcut.swagger.json --destination ./schema/shortcut.swagger.json && yarn validate:schema",
+    "validate:schema": "swagger-cli validate ./schema/shortcut.swagger.json",
     "lint": "eslint 'src/**/*.{js,ts}'",
     "test": "jest",
-    "prepublishOnly": "yarn build"
+    "format": "prettier --write -l *.{json,md,prettierrc} '{src,scripts}/**/*.ts'",
+    "format:check": "prettier --check *.{json,md,prettierrc} '{src,scripts}/**/*.ts'"
   },
   "dependencies": {
-    "axios": "^0.21.3",
+    "axios": "^1.5.0",
     "form-data": "^4.0.0"
   },
   "devDependencies": {
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.10.3",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jest": "^24.5.2",
-    "eslint-plugin-prettier": "^4.0.0",
-    "jest": "^27.2.4",
-    "jscodeshift": "^0.13.0",
+    "@types/jest": "^29.5.4",
+    "@types/jscodeshift": "0.11.6",
+    "@types/node": "^20.5.7",
+    "@typescript-eslint/eslint-plugin": "^6.5.0",
+    "@typescript-eslint/parser": "^6.5.0",
+    "eslint": "^8.48.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-jest": "^27.2.3",
+    "eslint-plugin-prettier": "^5.0.0",
+    "jest": "^29.6.4",
+    "jscodeshift": "^0.15.0",
     "node-wget": "^0.4.3",
-    "prettier": "2.4.1",
-    "rimraf": "^3.0.2",
+    "prettier": "^3.0.3",
+    "rimraf": "^5.0.1",
     "swagger-cli": "^4.0.4",
-    "swagger-typescript-api": "9.2.0",
-    "ts-jest": "^27.0.5",
-    "typedoc": "0.22.5",
-    "typedoc-plugin-merge-modules": "3.0.2",
-    "typescript": "4.4.3"
+    "swagger-typescript-api": "^13.0.3",
+    "typedoc": "0.25.0",
+    "typedoc-plugin-merge-modules": "5.1.0",
+    "typescript": "^5.2.2",
+    "ts-jest": "^29.1.1"
   },
   "jest": {
     "preset": "ts-jest",
